+++ swan/dataset/dgl_graph_data.py
@@ -22,7 +22,11 @@ PathLike = Union[str, Path]
 def collate_fn(samples):
     """Aggregate graphs."""
     graphs, y = map(list, zip(*samples))
-    return dgl.batch(graphs), torch.cat(y).unsqueeze(-1)
+    batch = dgl.batch(graphs)
+    if not any(item is None for item in y):
+        return batch, torch.cat(y).unsqueeze(-1)
+    else:
+        return batch, None
 
 
 def dgl_data_loader(*args, **kwargs):
@@ -53,6 +57,8 @@ class DGLGraphData(SwanGraphData):
         optimize_molecule
             Perform a molecular optimization using a force field.
         """
+        self.graph_creator = create_molecular_dgl_graph
+
         super().__init__(
             data_path, properties=properties, sanitize=sanitize,
             file_geometries=file_geometries, optimize_molecule=optimize_molecule)
@@ -62,19 +68,6 @@ class DGLGraphData(SwanGraphData):
 
         # define the loader type
         self.data_loader_fun = dgl_data_loader
-
-    def compute_graph(self) -> List[dgl.DGLGraph]:
-        """compute the graphs in advance."""
-        # create the graphs
-        molecular_graphs = []
-        for idx in range(len(self.labels)):
-            gm = create_molecular_dgl_graph(
-                self.dataframe["molecules"][idx],
-                self.dataframe["positions"][idx],
-                labels=self.labels[idx])
-            molecular_graphs.append(gm)
-
-        return molecular_graphs
 
     def get_item(self, batch_data: List[torch.Tensor]) -> Tuple[torch.Tensor, torch.Tensor]:
         """get the data/ground truth of a minibatch
@@ -106,5 +99,6 @@ class DGLGraphDataset(Dataset):
 
     def __getitem__(self, idx: int) -> Tuple[dgl.DGLGraph, torch.Tensor]:
         """Return the idx dataset element."""
+        label = None if len(self.labels) == 0 else self.labels[idx]
 
-        return self.molecular_graphs[idx], self.labels[idx]
+        return self.molecular_graphs[idx], label